refactor(sync): extract round-completion push to SyncEngine extension
SyncEngine.swift grew to 642 lines after Story 12.2 — over the 600-line
SwiftLint threshold. Move pushRoundCompletion and didRecentlyPushCompletion
into SyncEngine+RoundCompletion.swift. Main file is now 504 lines.

Relaxes two members from `private` to module-internal so the extension
file can reach them: `cloudKitClient` and `fetchAllMetadata()`.
